Add getWealthGrowthChange() and change float to int

+++ src/unsw/gloriaromanus/component/TaxLevel.java
@@ -1,15 +1,17 @@
 package unsw.gloriaromanus.component;
 
 public interface TaxLevel {
-    
+
     /**
      * Returns tax rate of corresponding tax level
+     * 
      * @return tax rate
      */
-    public float getTaxRate();
+    public int getTaxRate();
 
     /**
      * Returns wealth growth rate change of corresponding tax level
+     * 
      * @return wealth growth rate change
      */
     public int getWealthGrowthChange();
